Move recorder to file
